Adding ghostHunter support
Since Tapir/Tapirius seems to have permanently broken, I'm submitting a
pull request to have ghostHunter be included as the search engine of
Vapor. I run it on my own blog
[classicyuppie.com](http://classicyuppie.com) with success and the key
benefit above other search capabilities is that it offers real-time
search as you type, culling the results to a list of posts/pages that
actually useful.